Second part of checkmate mechanic
After king tile is found, I iterate each piece of the enemy board to see if any is a possible Check. If found, an object with the piece and the possible moves are return.

With that piece, i ll have to understand which tiles represent the move, so it can be block by my other pieces